add cache info modal and take animation settings out from SettingMenuPage.tsx
